changing names in output package, adding abstract class and editing gui

+++ sales-commissions-application/src/data/ReceiptManager.java
@@ -59,7 +59,7 @@ public class  ReceiptManager{
 		return sumItems;
 	}
 
-	public double calculateCommission(int totalSales) {
+	public double calculateCommission(double totalSales) {
 		if (totalSales > 6000 && totalSales <= 10000) {
 			return 0.1 * (totalSales - 6000);
 		}
